Test driver script for TravisCI
Adding test driver script for travis-ci testing.

This will get called by Travis-CI when the BSSw repository
gets changed.  Travis-CI will clone the scripts repository
into a temp directory, ___Scripts and runs this.

There's a lot of extra information in this script that we
probably don't need -- I was just putting in printouts to
see what a lot of the variables, etc. are thta Travis-CI
sets.