Bug: fixed a bug in the tester
A traceback happened when a deleted or moved file was in the change-set
because the tester tried to open it.

This is due to our diff file being created using `git diff --name-only`
which just prints a filename per line.

For now, it's ok to assume that if the file is gone then there's nothing
to test anyhow so we just catch the IOError exception and move on.

A better fix will be to use `git diff --name-status` which prints out
more information about what happened to a given file and parse that.