Bugfix: Badly formed lines in metadata handline.
Addresses PR #244 from the BSSW repository.
https://github.com/betterscientificsoftware/betterscientificsoftware.github.io/pull/344

An empty line in the metadata caused an exception.
- The fix is to just skip empty lines in metadata

Metadata lines with no ":" and thus not separable into a `<key>:<value>`
style pair are still problematic, but I'm adding in some exception
handling to that line splitter to at least print out the line that
caused the segfault and an error message.

Going forward we will need to determine how to handle "comment" lines in
the metadata.